Elide the GregTech 'You ran into  a serious Bug, if you have legitly acquired this Block, please report it immidietly' block name
Ideally should get the block name via the GregTech API, getDescription() on the TE;
Greg returns this message instead since the block information is not available
from the ID and metadata alone. But we get a fairly accurate name from the IInventory
getInvName(), and this message takes up a large portion of the screen, so don't show it.